# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Supernova API
  description: |
    REST API for the Supernova task management application.

    ## Overview
    Supernova is a task app where your lock screen wallpaper IS your task list,
    showing you the ONE thing to do right now.

    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    - **Access Token**: 7-day expiry, used for API access
    - **Refresh Token**: 30-day expiry, used to obtain new access tokens
    - **Wallpaper Token**: 365-day expiry, limited scope for widget access

    ## Rate Limiting
    - Global: 100 req/min
    - Auth endpoints: 10 req/min
    - Wallpaper: 20 req/hour
    - Task creation: 30 req/min

  version: 1.0.0
  contact:
    name: Supernova Support
    url: https://github.com/supernova-app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.supernova.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/user:
    get:
      tags:
        - User
      summary: Get user profile
      description: Retrieve current user's profile and settings
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - User
      summary: Update user settings
      description: Update user profile settings. Invalidates wallpaper cache.
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - User
      summary: Delete user account
      description: Delete user account and all associated data (GDPR compliance)
      operationId: deleteUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Account deleted successfully. All associated data has been removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/user/done-for-today:
    post:
      tags:
        - User
      summary: Mark done for today
      description: Mark user as done for the day (shows celebration wallpaper)
      operationId: markDoneForToday
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Marked as done
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  doneForToday:
                    type: boolean
                    example: true
                  doneAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # TASK ENDPOINTS
  # ============================================
  /api/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: |
        Get all active tasks for user, sorted by priority score.
        Excludes completed tasks and tasks snoozed until future.
      operationId: listTasks
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: location
          in: query
          description: Context for location-based prioritization
          schema:
            type: string
        - name: time
          in: query
          description: Time context for prioritization
          schema:
            type: string
            enum: [morning, afternoon, evening]
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Tasks
      summary: Create task
      description: |
        Create a new task with natural language parsing.
        Supports patterns like "Buy milk tomorrow at 3pm in 30m"
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              natural_language:
                summary: Natural language input
                value:
                  rawTitle: Buy groceries tomorrow at 3pm in 30m
              structured:
                summary: Structured input
                value:
                  title: Buy groceries
                  priority: 2
                  context_location: grocery_store
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Parse error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Tasks
      summary: Delete all tasks
      description: Delete all tasks for user
      operationId: deleteAllTasks
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: All tasks deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: All tasks deleted
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get a single task by ID
      operationId: getTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Tasks
      summary: Update task
      description: Update task fields. Sets completed_at/archived_at automatically.
      operationId: updateTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a single task
      operationId: deleteTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/tasks/{taskId}/snooze:
    post:
      tags:
        - Tasks
      summary: Snooze task
      description: Snooze task until tomorrow
      operationId: snoozeTask
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/taskId'
      responses:
        '200':
          description: Task snoozed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================
  # SYNC ENDPOINTS
  # ============================================
  /api/sync:
    post:
      tags:
        - Sync
      summary: Sync changes
      description: |
        Synchronize offline changes with server using last-write-wins conflict resolution.
        Returns all tasks modified since lastSyncAt.
      operationId: syncChanges
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '200':
          description: Sync complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/wallpaper:
    get:
      tags:
        - Wallpaper
      summary: Generate wallpaper
      description: |
        Generate and return wallpaper PNG image.
        Cached for 1 hour, invalidated on task/settings changes.
      operationId: getWallpaper
      security:
        - bearerAuth: []
      parameters:
        - name: theme
          in: query
          description: Wallpaper theme
          schema:
            type: string
            enum: [cosmic, ocean, fantasy]
            default: cosmic
        - name: resolution
          in: query
          description: Screen resolution (WxH)
          schema:
            type: string
            pattern: '^\d+x\d+$'
            default: 1080x1920
          example: 1080x1920
        - name: enhanced
          in: query
          description: Use enhanced generator
          schema:
            type: boolean
            default: true
        - name: timestamp
          in: query
          description: Animation frame timestamp (skips cache)
          schema:
            type: integer
      responses:
        '200':
          description: Wallpaper PNG image
          headers:
            X-Cache:
              description: Cache hit status
              schema:
                type: string
                enum: [HIT, MISS]
            X-Fallback:
              description: Whether fallback was used
              schema:
                type: string
          content:
            image/png:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # UTILITY ENDPOINTS
  # ============================================
...
  /api/done-for-today:
    post:
      tags:
        - Utility
      summary: Mark done for today (legacy)
      description: Legacy endpoint for marking done for today
      operationId: markDoneForTodayLegacy
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Marked as done
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    taskId:
      name: taskId
      in: path
      required: true
      description: Task UUID
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              message: Authentication required
              code: AUTHENTICATION_ERROR
...